+++ crates/ruma-client-api/src/keys/upload_signatures.rs
@@ -45,6 +45,7 @@ pub mod v3 {
     #[derive(Default)]
     pub struct Response {
         /// Signature processing failures.
+        #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
         pub failures: BTreeMap<OwnedUserId, BTreeMap<String, Failure>>,
     }
 
@@ -117,14 +118,17 @@ pub mod v3 {
         _Custom(PrivOwnedStr),
     }
 
-    #[cfg(all(test, feature = "client", feature = "compat-upload-signatures"))]
+    #[cfg(all(test, feature = "client"))]
     mod tests {
-        use ruma_common::user_id;
+        use super::ResponseBody;
 
-        use super::{FailureErrorCode, ResponseBody};
-
+        #[cfg(feature = "compat-upload-signatures")]
         #[test]
         fn deserialize_synapse_response() {
+            use ruma_common::user_id;
+
+            use super::FailureErrorCode;
+
             const JSON: &str = r#"{
                 "failures": {
                     "@richvdh:sw1v.org": {
@@ -141,6 +145,14 @@ pub mod v3 {
             let failure = &parsed.failures[user_id!("@richvdh:sw1v.org")]["EOZDSWJVGZ"];
             assert_eq!(failure.errcode, FailureErrorCode::InvalidSignature);
             assert_eq!(failure.error, "400: Invalid signature");
+        }
+
+        #[test]
+        fn deserialize_empty_response() {
+            const JSON: &str = r#"{}"#;
+
+            let _parsed: ResponseBody = serde_json::from_str(JSON)
+                .expect("We should be able to deserialize an empty keys/signatures/upload");
         }
     }
 }