+++ crates/ruma-client-api/src/account/deactivate.rs
@@ -39,6 +39,13 @@ pub mod v3 {
         /// identifier.
         #[serde(skip_serializing_if = "Option::is_none")]
         pub id_server: Option<String>,
+
+        /// Whether the user would like their content to be erased as much as possible from the
+        /// server.
+        ///
+        /// Defaults to `false`.
+        #[serde(default, skip_serializing_if = "ruma_common::serde::is_default")]
+        pub erase: bool,
     }
 
     /// Response type for the `deactivate` endpoint.